# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " stores API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8080",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/stores": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Récupère tous les magasins",
                "description": "Récupère la liste des magasins avec des filtres et options de tri.",
                "operationId": "98d4ef27c7e1f04920fad64db308cdc7",
                "parameters": [
                    {
                        "name": "filter_by",
                        "in": "query",
                        "description": "Nom de la colonne sur laquelle appliquer le filtre (name, adress, zipcode, city, country, phone_number).",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "adress",
                                "zipcode",
                                "city",
                                "country",
                                "phone_number"
                            ]
                        }
                    },
                    {
                        "name": "filter_value",
                        "in": "query",
                        "description": "Valeur à rechercher dans la colonne spécifiée par filter_by.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Nom du champ pour trier les magasins.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Ordre de tri (ASC ou DESC).",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ASC",
                                "DESC"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des magasins récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "adress": {
                                                "type": "string"
                                            },
                                            "zipcode": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des magasins"
                    }
                }
            }
        },
        "/stores/{id}": {
            "get": {
                "tags": [
                    "Stores"
                ],
                "summary": "Récupère un magasin spécifique",
                "description": "Récupère un magasin par son ID.",
                "operationId": "f4dd45d787db0214dde1d8883842f86a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du magasin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Magasin récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "adress": {
                                            "type": "string"
                                        },
                                        "zipcode": {
                                            "type": "string"
                                        },
                                        "city": {
                                            "type": "string"
                                        },
                                        "country": {
                                            "type": "string"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Magasin non trouvé"
                    },
                    "500": {
                        "description": "Erreur lors de la récupération du magasin"
                    }
                }
            }
        },
        "/stores/edit/{id}": {
            "put": {
                "tags": [
                    "Stores"
                ],
                "summary": "Met à jour un magasin",
                "description": "Met à jour un magasin existant par son ID.",
                "operationId": "625e1f0b204b83a753ab153be1dcbd1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du magasin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "adress",
                                    "zipcode",
                                    "city",
                                    "country",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "zipcode": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Magasin mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides ou aucune modification"
                    },
                    "404": {
                        "description": "Magasin non trouvé"
                    },
                    "500": {
                        "description": "Erreur lors de la mise à jour du magasin"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/stores/create": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Crée un nouveau magasin",
                "description": "Crée un magasin avec les données fournies.",
                "operationId": "aa7e9a2cf0b855f535d46e3430ac466a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "adress",
                                    "zipcode",
                                    "city",
                                    "country",
                                    "phone_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "zipcode": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Magasin créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Champs requis manquants ou données invalides"
                    },
                    "500": {
                        "description": "Erreur lors de la création du magasin"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
        "/stores/delete/{id}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Supprime un magasin",
                "description": "Supprime un magasin par son ID.",
                "operationId": "f453cc92d21818ea586a930f6e9a0497",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du magasin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Magasin supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID invalide"
                    },
                    "404": {
                        "description": "Magasin non trouvé"
                    },
                    "500": {
                        "description": "Erreur lors de la suppression du magasin"
                    }
                },
                "security": [
                    {
                        "BearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}